workers/tkr_ibm_kobe/pyproject.toml
[project]
name = "tkr-ibm-kobe"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pytket-qiskit>=0.72.0",
    "tierkreis",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "ruff>=0.14.3",
]
